[Bronze V] Title: 사칙연산, Time: 208 ms, Memory: 17720 KB -BaekjoonHub

+++ 백준/Bronze/10869. 사칙연산/사칙연산.java
@@ -2,12 +2,12 @@ import java.util.Scanner;
 class Main{
     public static void main(String[] args){
         Scanner ab = new Scanner(System.in);
-        double a = ab.nextDouble();
-        double b = ab.nextDouble();
-        System.out.println((int)(a+b));
-        System.out.println((int)(a-b));
-        System.out.println((int)(a*b));
-        System.out.println((int)(a/b));
-        System.out.println((int)(a%b));
+        int a = ab.nextInt();
+        int b = ab.nextInt();
+        System.out.println(a+b);
+        System.out.println(a-b);
+        System.out.println(a*b);
+        System.out.println(a/b);
+        System.out.println(a%b);
     }
 }
